api/requirements.txt
flask==1.1.1
flask_cors==3.0.9
flask_limiter==1.4
python-chess==1.999
python-dotenv==0.15.0